+++ .gitignore
@@ -1,2 +1,3 @@
 loc
-a.out
+*.o
+*.a
