+++ files.h
@@ -1,0 +1,7 @@
+#ifndef FILES_H
+#define FILES_H
+
+long int file_size(char *path);
+int count_lines(char *path);
+
+#endif
